only make user available to add, update, delete course if this course is owned by the user

+++ controllers/courses.js
@@ -28,11 +28,21 @@ export const getCourse = asyncHandler(async (req, res, next) => {
 
 export const addCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId
+  req.body.user = req.user.id
 
   const bootcamp = await Bootcamp.findById(req.params.bootcampId)
 
   if (!bootcamp) {
     return next(new ErrorResponse(`No bootcamp with the id of ${req.params.bootcampId}`), 404)
+  }
+
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+        401
+      )
+    )
   }
 
   const course = await Course.create(req.body)
@@ -45,6 +55,15 @@ export const updateCourse = asyncHandler(async (req, res, next) => {
 
   if (!course) {
     return next(new ErrorResponse(`No course with the id of ${req.params.id}`), 404)
+  }
+
+  if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update a course ${req.params.id}`,
+        401
+      )
+    )
   }
 
   course = await Course.findByIdAndUpdate(req.params.id, req.body, {
@@ -60,6 +79,15 @@ export const deleteCourse = asyncHandler(async (req, res, next) => {
 
   if (!course) {
     return next(new ErrorResponse(`No course with the id of ${req.params.id}`), 404)
+  }
+
+  if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete course ${req.params.id}`,
+        401
+      )
+    )
   }
 
   await course.deleteOne()
